requirements.txt
functions-framework==3.*
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
idna==3.4
pymongo==4.5.0
requests==2.31.0
urllib3==2.0.4
python-dotenv==1.0.0